pagination added with the routing

+++ controllers/admin.js
@@ -1,3 +1,7 @@
+const mongoose = require('mongoose');
+
+const  fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator/check')
 
 const Product = require('../models/product');
@@ -15,9 +19,26 @@ exports.getAddProduct = (req, res, next) => {
   
   exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const description = req.body.description;
     const price = req.body.price;
+
+    if(!image){
+      return res.status(422).render('admin/edit-product', {
+        pageTitle: 'Add Product',
+        path: '/admin/add-product',
+        editing: false,
+        hasError: true,
+        product: { 
+          title: title,          
+          price: price,
+          description: description
+        },
+        errorMessage: 'Attached file is not an image.',
+        validationErrors: []
+      });
+    }
+
     const errors =  validationResult(req);
 
     if(!errors.isEmpty()){
@@ -27,8 +48,7 @@ exports.getAddProduct = (req, res, next) => {
         editing: false,
         hasError: true,
         product: { 
-          title: title,
-          imageUrl: imageUrl,
+          title: title,          
           price: price,
           description: description
         },
@@ -36,6 +56,8 @@ exports.getAddProduct = (req, res, next) => {
         validationErrors: errors.array()
       });
     }
+
+    const imageUrl = image.path;
 
     const product = new Product({
       title: title, 
@@ -90,7 +112,7 @@ exports.getAddProduct = (req, res, next) => {
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
     const updatedDesc = req.body.description;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
 
     const errors =  validationResult(req);
 
@@ -120,7 +142,10 @@ exports.getAddProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if(image){
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path; 
+      }
       return product.save().then(result => {
         res.redirect('/admin/products');
         console.log('Product Updated Successfully');
@@ -151,8 +176,13 @@ exports.getAddProduct = (req, res, next) => {
 
   exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
-    .then(() => {
+    Product.findById(prodId).then(product => {
+       if(!product){
+         return next(new Error('Product Not Found'));
+       }
+       fileHelper.deleteFile(product.imageUrl);
+       return Product.deleteOne({_id: prodId, userId: req.user._id});
+    }).then(() => {
       res.redirect('/admin/products');
       console.log('Product Deleted');
     })
@@ -160,5 +190,6 @@ exports.getAddProduct = (req, res, next) => {
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
-    });
+    });    
+    
   }
